Append MultiCollectorTest to TestMultiCollector. (#1745)
